chore(refactor): rename `context/modals` to `controllers/modals`
