+++ src/context/DO_NOT_TOUCH.md
@@ -1,0 +1,2 @@
+hello do not touch `intermediate` or `revoltjs` folders
+they are being rewritten
