+++ src/pages/settings/GenericSettings.tsx
@@ -13,9 +13,8 @@ import { isTouchscreenDevice } from "../../lib/isTouchscreenDevice";
 
 import { useApplicationState } from "../../mobx/State";
 
-import { modalController } from "../../context/modals";
-
 import ButtonItem from "../../components/navigation/items/ButtonItem";
+import { modalController } from "../../controllers/modals/ModalController";
 
 interface Props {
     pages: {
